Fix having to detach the user identity
In the `CompanyUserAdapter` and `UserAdapter`, their respective
user entity was detached from Doctrine's `EntityManager`, this meant
that updates that involved the users would fail due to discovery of
a "new" managed version of that entity.

This required a bit of hacky code in the `LoginAttempt` service to
fix this.

This is no longer required, as there is no reason to detach the user
entities from the `EntityManager` in the first place.