Enable marking course documents as scanned
Scanned course documents are often of lower quality than actual PDFs,
this results in pixelated downloads where somethings certain parts
of the original scan are removed due to the flattening process.

By marking a course document as being scanned, it can be downloaded
at a higher quality. It should be noted that setting this option for
normal PDFs will significantly affect the download size and speed,
likely resulting in no file being downloaded.