Make `BaseMapper` a template and fix related typing issues
By making the `BaseMapper` a template it will be easier to see what
a mapper actually returns. This should prevent us from having to
use statements involving `/** @var SomeModel $variable */`.

This also fixes almost half of the type issues that we had according
to Psalm. Not everything can and should be fixed but this helps.