Switch to `declare(strict_types=1)` and start enforcing it
This change will help preventing type confusion mistakes. As PHP does
not have a compiler to do type checks, this is the closest we can get
together with our existing inspection tools (PHP_CS and Psalm).

In this comment all of the changes necessary to make this work are
added. These mostly include fixes to remove type juggling.