Change `HEALTHCHECK` for MariaDB
MariaDB v11.* saw the removal of all `mysql*` commands to better
distinguish it from MySQL. However, this was of course not (clearly)
communicated resulting in breaking many healthchecks.

On top of that, their own healthcheck contained a bug that required
the usage of an empty root password. This was fixed a few days ago
so now everything should work again with the changes in this commit.