Add language aware route stack to allow language specific links
This route stack can be used to parse and handle the routing of
links which include a language identifier, for example, `en` and
`nl`.

The code to determine the preferred language of the visitor has
been removed from `onBootstrap()` to reduce its overhead. It has
been moved to the new route stack to handle everything during the
route matching phase.

A listener has been added for `MvcEvent::EVENT_ROUTE` to ensure
that we can get the language from the route matching into the
actual translators and relevant i18n code.

The new route stack also enables the automatic inclusion of the
preferred language in the generation of links through the `url()`
view helper. Manually changing the `language` parameter is fully
supported too.

---

There is currently no support for translatable routes even though
we extend the `TranslatorAwareTreeRouteStack`. On top of that,
future changes will add support for language specific routes for
custom pages.