Fix 500 on 404 for unroutable URLs
When a URL did not fit within the parameters of the route stack an
exception would be thrown when trying to create the language
switching URLs.

Because it does not make sense to switch the language on these
unroutable URLs (there is no point to recovery) we link to the
homepage instead.