Move `Page`s to `FrontpageLocalisedText` and i18n routes
With this change, all fields of the custom `Page`s are now using the
`FrontpageLocalisedText` implementation. This means that together with
the `LanguageAwareTreeRouteStack` we can now have different routes
for the Dutch and English versions of a custom page.

To make this possible, the `Page` form has been extended to accept
pairs for the routes instead of single values. To make sure that you
can never have two distinct pages with the same routes, the validation
of the form has been extended.

On creation of new pages the routes are checked against all known
routes from the database. If a page is edited, it is determined whether
the route needs to be checked (against the database).

However, this means that you cannot use the Dutch route when viewing
the website in English and vice-versa.