GEWISDB sync:truncate -> INSERT/DELETE/UPDATE
With this commit, the database sync no longer truncates all data in the table. The TRUNCATE statement is removed (which had an implicit COMMIT, causing the website to be unavailable for a short period of time during sync and is replaced by INSERT ... ON DUPLICATE KEY UPDATE in combination with a DELETE FROM statement that removes no longer existing entries. Foreign key checks are still disabled because the foreign keys do not have any cascading settings.