Add better error handling and earlier `exit()`s to GEWISDB sync
Changes important string concatenations to `sprintf` to make it
easier to see what actually happens. To enable this the `catch`
blocks haveto be able to catch `Error`s (internal PHP errors)
next to `Exception`s.

This also changes the DSNs to take the `port`s and `charset`s
of the databases into account and moves `setAttribute()` to the
constructor. Technically, `PDO::ATTR_ERRMODE` defaults to
`PDO::ERRMODE_EXCEPTION`, however, being setting it explicitly
should make it more obvious.

Added `ROLLBACK` statement if a top-level error occurs, inserts
and removals are caught separately. Moved the restore of the
foreign key constraints to a `finally` block to ensure it is
always executed (as long as the connection is still working).