Allow default-to-open dropdowns to be closed on smaller screens
Achieved by removed the styling that forced the dropdown to always
be shown (regardless of state) and adding `.open` by default.

To make sure that the dropdown is open again after uncollapsing the
navbar, we add `.open` back on the `shown.bs.collapse` event. This
will result in the user seeing the dropdown open but there is no
better way to do this (`show.bs.collapse` is too early).