Fix aspect ratio for organ cover image
The viewable aspect ratio was always 16:4, however, the cropping
tool used an aspect ratio of 16:5. The cropping tool now has the
correct aspect ratio of 16:4, on top of that the explanation has
been changed to explicitly state the aspect ratio.

On top of that, the `aspect-ratio` SCSS mixin has been removed.