Move SCSS source files out of webroot
This moves the SCSS source files out of the webroot to ensure that
they are not accessible from the actual website. Furthermore, the
Bootstrap and FontAwesome source files are no longer kept in version
control. This should make it easier to perform dependency upgrades
without creating massive changes.

To facilitate this the `Dockerfile`s had to be changed. To make sure
that the relevant files are copied just in time for compilation of
the CSS files. This also means that we dropped
`slowprog/composer-copy-file`, which we used to copy the files after
a Composer install or update.

Furthermore, this removes the dependency on an external request to
be able to install Composer. The Composer binary now comes directly
from the Composer Docker image. As such, the `composer.phar` has
also been removed from version control.