Change wording of privacy widget to prevent confusion
GEWIS uses analytical cookies and NOT tracking cookies to:
- track the number of visitors to our web pages,
- track the length of time each visitor spends on our web pages,
- view the order in which a visitor visits the various pages of our website,
- assess which parts of our site require adjustment, and
- optimise the website.

As such, we improve the wording to not mention tracking as it may
confuse the visitor of the website.

The reasons above are valid reasons by law to make the analytical
cookies mandatory (in contrast to tracking cookies which should
be opt-in). However, we still want to provide the visitor with the
opportunity to disable them.