Move activity creation to admin section
To ensure that there is enough space for the new Markdown editor
the activity creation page has been moved to the admin section of
the website.

Nothing changes in terms of functionality.

Unfortunately, the indent changes make the old and new file too
dissimilar for `git` to recognise it as a move instead of deletion
and creation.