Move `Activity`, `NewsItem`, and `OrganInformation` editors to Markdown
For `Activity` descriptions Markdown was already the standard but
the editor nor the parser enforces this. This can lead to people
breaking the layout of the page when their HTML is not properly
structured and/or formatted.

To make sure that this is not possible, the editor now ONLY supports
Markdown. Technically it is the GitHub Flavour of Markdown.

Additionally, this removes the old and outdated editor which was
broken since its introduction (the preview functionality was never
implemented).

`NewsItem`s and `OrganInformation` long text/description also have no
need for custom and/or complex HTML formatting, so that editor has
also been switched to only support GFM.