Switch to `league/commonmark` for Markdown parsing
`maglnet/maglmarkdown` appears to be abandoned, as such it has been
replaced with `league/commonmark`. `league/commonmark` was actually
one of the Markdown converts in `maglnet/maglmarkdown`.

This is essentially a drop-in replacement with the addition of the
`Markdown` view helper. Configuration ensures that GitHub Flavoured
Markdown is supported. Additionally, HTML content is escaped and
not stripped. Unsafe URLs are not allowed. External URLs are
automagically changed to open in a new tab, with the appropriate
`rel` attribute.