Switch activity subscription from well to modal
To make it easier to subscribe to activities, the subscription form
is now a modal. This prevents the page from scrolling and allows
for some magic auto re-openings of the modal if there is an error.

This also changes the states the activity subscription button can
be in to ensure that it is easier to see what will happen.