Switch company and job descriptions to special Markdown editor
Utilises the new MarkdownCompanyEditor that also allows photos to
be uploaded and videos to be inserted.

Files from each company are now stored in a company specific folder
to prevent these files from overwriting other files (either those
from GEWIS itself or other companies).

IMPORTANT: files are stored based on the ID of the company not the
slug. This is because the slug can change and the ID should not
change.