Fix ordering of functions for committees and organs
Ordering is now guaranteed to follow:
- Chair
- Secretary
- Treasurer
- Vice-Chair

If someone has multiple functions then the highest function is used
for the actual ordering. So someone who is both secretary and vice-
chair will be above someone who is treasurer. Likewise, if someone
is treasurer and vice-chair they will still be ranked below chair
and secretary.