Improved detailed organ view
The ordered subdecisions in an `Organ` are now also sorted by their
individual subdecision number, this is to ensure correct ordering in
the list.

Additionally, the formatting has changed to show in which meeting the
decision was made (with clickable link) and the actual (sub)decisions
can now also be linked.