Add `activity:gdpr:delete-old-signups` command to remove old `Signup`s
In accordance with GEWIS' privacy policy, `Signup`s for an `Activity`
that ended 5 years ago should be removed. Previously, I did this
manually, however, as I am doing less and less for GEWISWEB this
process should be automated.

As such, this command is the first in a series of command that will
automate some GDPR related tasks. All these tasks are expected to
be executed between 03:00 and 05:00 each day.