Add automatic birthday balloons
For the past few years I have manually added the balloons to the website on the
28th of June, however, it is better to just have this implemented so I do not
have to do this anymore.

For GEWIS' 40th birthday it was decided that 40 balloons was optimal for both
small and large screens. This should not be changed without careful
consideration as it may negatively affect the usability of the website.

---

Removal of entry in `.gitignore` came from GH-1790 to ensure that the SCSS files
are tracked.