feat: refactor authentication exception flow
Introduces better support for detecting _who_ is (or is not) signed in and as a
result we can distinguish between 401 and 403 errors. As a result, we can now
also show a user they may have to switch between being logged in as a member or
a company representative.

This does not yet change how we handle authentication for people, this will
follow in later commits. However, this has already changed authentication for
the API. From now on, API clients should authenticate by using the
`Authorization` header using the `Bearer` option with an access token. The old
way of authenticating will continue to work for now but will be removed in the
future.

Co-Authored-By: Rink <rinkp@users.noreply.github.com>