chore(seeder): add Member and User seeders
The data fixtures can be loaded into the database using the
`application:fixtures:load` command. All existing records are `TRUNCATE`d from
the database to ensure a clean start.

This also fixes some inconsistencies in the (sub)decision model with GEWISDB,
somehow the possibility for these to be `null` got lost somewhere (and fixes for
initialisation of `Collection`s).

---

Unfortunately, adding the data fixtures for (sub)decisions has proved to be
quite difficult. As such, these have been removed. The WIP can be found in
https://github.com/GEWIS/gewisweb/pull/1913.

There is an issue with the "hydration" of the entities when they are added to
the database. I have not seen this issue in GEWISDB, but the cause appears to be
the usage of `BackedEnum`s as part of a composite key (which forms the
foundation for our (sub)decision entities and relations).

Either the enum cannot be cast to string while being saved to the database. Or
when using custom mapping types (see the PR mentioned) above the value cannot
be properly restored from the database. The latter can then also be fixed with
another patch for ORM (see https://github.com/GEWIS/orm/commit/8031547),
however, this may break other things. This patch can probably also be applied in
reverse, such that we do not need the custom mapping types. However, this has
not (yet) been tested.

As such, this has to be investigated more and potentially a bug report must be
submitted to Doctrine ORM to get this fixed.