feat: improved cards for bodies, companies, and jobs
Started out as a rewrite for the broken cards used to display BM and GMM bodies.
I believe I broke their spacing many years ago when refactoring something else,
but I never had the need to fix it (as people stopped complaining).

The cards for companies always worked but looking back at the Sass code, they
were just interfering with each other in which the body overview lost.

Cards are now unified and follow the newer Boostrap styling more closely.

---

Intention was to fix the styling for BM/GMM bodies where the abbreviation of the
body was not visible due to the usage of picture with a white(-ish) background.
This has been resolved by moving all text back into the body of the card instead
of on top of the cover image. Sizing and grid layout are handled through CSS.

If there is no cover image, the GEWIS logo is now shown (similar to how this is
done in the Photo module).