feat: improved activity overview
Gives the description of an activity more space (but also clamping it to at most
5 lines) by removing all event details on the right.

Event details are now directly under the activity title (for date and time) or
at the bottom (for costs and location). The date and time format has been
condensed and a helper "for another X" / "in X" has been added to see how long
an activity will still run or when it will start.

Also allows links to be clicked in the description.