fix: time diff calculation
I used `ceil` instead of `floor` for weeks to help with underestimation of the
time till an activity (or how long an activity would still run for). However,
this resulted in a lot of confusion as it was possible for an activity to
"happen" in two weeks while it would have been 1 week and 1 day.

As such, the formatter now displays with multiple units if possible. But only
for years, months, weeks, and days OR hours and minutes OR minutes and seconds.

--

For Dutch, this also changes `jaren` to `jaar` and `uren` to `uur`.