chore: update dependencies
We will not be moving to PHP 8.4 for quite some time (also something something
Laminas Service Manager v4).

The dependency updates address some critical security issues (mostly related to
DoS attacks). This includes changes to the Commonmark config to limit the number
of delimiters per line.

CKEditor 5 has not been upgraded, as we are already on the last version that
support our current way of building it (from v42 onwards, they want you to
include the _whole_ editor and selectively select plugins from that).