fix: sizing of company logos in cards
Refactor of the card component removed the height restriction because the logos
are not the same size (nor share an aspect ratio like cards for BM/GMM bodies).

While I think that companies should provide proper logos (not ones with hundreds
of white pixels as margin on each side), the board has asked to revert to same
sizing to prevent certain companies from standing out in the overview.