fix: styling of activity and photo album year selectors
Both now use `nav-tabs` and when a year is selected from the dropdown, both the
dropdown _and_ the year are now marked as active.

This is very minor but should make it easier to see what you are looking at as
the year is not necessary displayed on the page itself.