fix: time diff calculation for activities
An error in the logic existed due to a lack of documentation. As a result, the
estimated times were off by up to 30 days.

As an example, an activity that takes place on February 14, 2025 at 17:00 would
state that this was in "1 month and 5 days" (at the time of writing this commit
message). We all know that this is not the case, because that would actually be
January 21, 2025.

The `d` property of a `DateInterval` is now used directly instead of doing our
own calculations with the `days` property (_all_ days between both dates, not
just the difference after substracting years and months).