feat: show when sign-up list opens/closes in activity overview
If an activity has one or more sign-up lists, it will display when this sign-up
list will open, and if the sign-up list is open when it will close.

If the sign-up list closes within a week, it is displayed with a subtle orange
colour. If the sign-up list closes within a day, the text is made red.