chore: optimise GEWIS logo in error pages
When I created this using a base64-encoded PNG URI was the easiest, however, the
URI is more than 30k characters. Doing it with the SVG makes it scale better and
is only 1.5k characters (unfortunately, the URL encoding is necessary). That is
a reduction of 95%.

---

The real reason for doing is is that I have been working on the custom error
pages for the CBC and needed the optimisations for that.