fix(photo): combination of week number and year being wrong
Week number is based on ISO 8601, which allows January 1 to fall on Mon-Thu for
that week to be the first week of the year. However, we cannot get the year from
the date itself, because if (for example for this year) that date still falls in
December the year is one year too low.