feat(activity): add Zettle as option for facilities
GEFLITST and Zettle are now "facilities" instead of "options" for an activity.

---

I had a lengthy discussion with the BAC on whether they would want to be
included in a similar way. In the end, we decided together that this was not
desirable for the BAC. This is because people mostly already (have to) contact
the BAC and this would then become more of a reminder (which is not necessary).

From AViCo I have unfortunately never had a response in the past 2 months,
should they still want to do so they are free to provide their own PR.