fix: GEWIS console splash being fully external
We host everything ourselves to ensure that it cannot be easily altered and we
have no issues with GDPR (similar to why we removed the Google-hosted fonts).
The former could have been fixed with subresource integrity hashes, but with the
privacy argument as well it is just easier to do it like this.

Similar changes have been added (or are pending) for Aurora and SudoSOS.