feat: maximum session age of 90 days
Sessions for which the user has opted to be remembered will now only be valid
for 90 days or till July 1st (whichever comes first).

The reasoning for this is because otherwise it would be possible to have never
ending sessions. In practice this was not possible, as I have been resetting all
sessions on July 1st and 2-3 times per year for the past few years.

Old behaviour where a created session is valid for 2 weeks is still in effect.
As such, if a user logs in but subsequently does nothing for the next two weeks
that session will naturally expire.