feat(activity): promoted ordering of signup lists
It may happen that an activity has more than one signup list, in certain cases
it may be preferred to use a different ordering than when the activity was made.

For example, when it is decided that an activity should be re-opened for signups
but you want to preserve the original signup list I have to manually create a
new signup list for this activity. To ensure that it is the first signup list in
the overview, I have to make manual changes in the database to fix the
positioning by altering the IDs of the signup lists (which is a pain).

Therefore, this introduces a structured approach to handle ordering directly
within the codebase. The functionality is *NOT* exposed to activity organisers.

Although more complex ordering (like considering signup list open/close times)
is possibly better, it is currently out-of-scope for this implementation (as I
am not entirely sure it is possible with Doctrine without have to create custom
queries).